Add files via upload
I'm not sure if I was supposed to do the Gaussian, but it was on his assignment, so I did it anyways. Can anyone check my calculations for the Gaussian? They are approximations to the teacher's work, but not exactly the same. I'm not sure if I should change it because I just followed the book's instructions. Also feel free to comment on my random number generator (chrono). I assume thats what he wanted, but if not I can change it.