Create QuadraticE_main.cpp
This will be the main.cpp file for the quadratic expression must have the QuadraticE.h + QuadraticE.cpp. 

Note: My main.cpp has been acting weird It's been giving me an error in line 1 so I need someone to check it for me, I think it's my menu but idk.